# HW3/openapi.yaml
swagger: "2.0"
info:
  description: "Homework description"
  version: "1.0.0"
  title: "LAB FE NodeJS HW #3"
  termsOfService: "http://swagger.io/terms/"
  contact:
    email: "Kyrylo_Yezhov@epam.com"
host: "localhost:8080"
basePath: "/"
...
schemes:
- "http"
paths:
  /api/users/me:
    get:
      tags:
      - "User"
      summary: "Get user's profile info"
      description: "User can request only his own profile info"
      operationId: "getProfileInfo"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Success"
          schema:
            type: object
            properties:
              user:
                $ref: '#/definitions/User'
        "400":
          description: "Bad request"
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: "Internal server error"
          schema:
            $ref: '#/definitions/Error'
      security:
      - jwt_token: []
    delete:
      tags:
      - "User"
      summary: "Delete user's profile"
      description: "User can delete only his own profile info"
      operationId: "deleteProfile"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
            description: "Success"
            schema:
                type: "object"
                properties:
                  message: 
                    type: string
                    example: "Profile deleted successfully"
        "400":
          description: "Bad request"
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: "Internal server error"
          schema:
            $ref: '#/definitions/Error'
      security:
      - jwt_token: []
  /api/users/me/password:
    patch:
      tags:
      - "User"
      summary: "Change user's password"
      description: "Change user's password"
      operationId: "changeProfilePassword"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "new_password"
        schema:
          type: "object"
          properties:
            oldPassword:
              type: string
            newPassword:
              type: string
      responses:
        "200":
            description: "Success"
            schema:
                type: "object"
                properties:
                  message: 
                    type: string
                    example: "Password changed successfully"
        "400":
          description: "Bad request"
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: "Internal server error"
          schema:
            $ref: '#/definitions/Error'
      security:
      - jwt_token: []
  /api/auth/register:
    post:
      tags:
      - "Auth"
      summary: "Register a new system user(Shipper or Driver)"
      description: "Create new profile"
      operationId: "createProfile"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "credentials"
        schema:
          $ref: "#/definitions/registrationCredentials"
      responses:
        "200":
            description: "Success"
            schema:
                type: "object"
                properties:
                  message: 
                    type: string
                    example: "Profile created successfully"
        "400":
          description: "Bad request"
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: "Internal server error"
          schema:
            $ref: '#/definitions/Error'
...
  /api/auth/forgot_password:
    post:
      tags:
      - "Auth"
      summary: "Forgot password"
      description: "(Optional) Forgot password option"
      operationId: "forgotPassword"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "credentials"
        schema:
          properties:
            email:
              type: "string"
              example: "kyrylo@gmail.com"
      responses:
        "200":
            description: "Success"
            schema:
                type: "object"
                properties:
                  message: 
                    type: string
                    example: "New password sent to your email address"
        "400":
          description: "Bad request"
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: "Internal server error"
          schema:
            $ref: '#/definitions/Error'
  /api/trucks:
    get:
      tags:
      - "Truck"
      summary: "Get user's trucks"
      description: "Retrieve the list of trucks for authorized user(abailable only for driver role)"
      operationId: "getUserTrucks"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      responses:
        "200":
          description: "Success"
          schema:
              type: "object"
              properties: 
                  trucks:
                      type: array
                      items:
                          $ref: "#/definitions/Truck"
        "400":
          description: "Bad request"
          schema:
            $ref: '#/definitions/Error'
        "500":
          description: "Internal server error"
          schema:
            $ref: '#/definitions/Error'
      security:
      - jwt_token: []
    post:
        tags:
        - "Truck"
        summary: "Add Truck for User"
        description: "Add Truck for User(abailable only for driver role)"
        operationId: "addUserTruck"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "body"
          name: "Truck payload"
          schema:
            properties:
              type:
                type: "string"
                enum: ["SPRINTER", "SMALL STRAIGHT", "LARGE STRAIGHT"]
                example: "SPRINTER"
        responses:
            "200":
                description: "Success"
                schema:
                    type: "object"
                    properties:
                      message: 
                        type: string
                        example: "Truck created successfully"
            "400":
                description: "Bad request"
                schema:
                    $ref: '#/definitions/Error'
            "500":
                description: "Internal server error"
                schema:
                    $ref: '#/definitions/Error'
        security:
        - jwt_token: []
  /api/trucks/{id}:
    get:
        tags:
        - "Truck"
        summary: "Get user's truck by id"
        description: "Get user's truck by id(abailable only for driver role)"
        operationId: "getUserTruckById"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "path"
          name: "id"
          description: "Truck id"
          required: true
          type: string
        responses:
            "200":
              description: "Success"
              schema:
                type: object
                properties:
                  truck:
                      $ref: '#/definitions/Truck'
            "400":
              description: "Bad request"
              schema:
                  $ref: '#/definitions/Error'
            "500":
              description: "Internal server error"
              schema:
                  $ref: '#/definitions/Error'
        security:
        - jwt_token: []
    put:
        tags:
        - "Truck"
        summary: "Update user's truck by id"
        description: "Update user's truck by id (abailable only for driver role)"
        operationId: "updateUserTruckById"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "path"
          name: "id"
          description: "Truck id"
          required: true
          type: string
        - in: "body"
          name: "Truck payload"
          schema:
            properties:
              type:
                type: "string"
                enum: ["SPRINTER", "SMALL STRAIGHT", "LARGE STRAIGHT"]
                example: "SPRINTER"
        responses:
            "200":
                description: "Success"
                schema:
                    type: "object"
                    properties:
                      message: 
                        type: string
                        example: "Truck details changed successfully"
            "400":
              description: "Bad request"
              schema:
                  $ref: '#/definitions/Error'
            "500":
              description: "Internal server error"
              schema:
                  $ref: '#/definitions/Error'
        security:
        - jwt_token: []
    delete:
        tags:
        - "Truck"
        summary: "Delete user's truck by id"
        description: "Delete user's truck by id (abailable only for driver role)"
        operationId: "deleteUserTruckById"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "path"
          name: "id"
          description: "Truck id"
          required: true
          type: string
        responses:
            "200":
                description: "Success"
                schema:
                    type: "object"
                    properties:
                      message: 
                        type: string
                        example: "Truck deleted successfully"
            "400":
              description: "Bad request"
              schema:
                  $ref: '#/definitions/Error'
            "500":
              description: "Internal server error"
              schema:
                  $ref: '#/definitions/Error'
        security:
        - jwt_token: []
  /api/trucks/{id}/assign:
    post:
        tags:
        - "Truck"
        summary: "Assign truck to user by id"
        description: "Assign truck to user by id (abailable only for driver role)"
        operationId: "assignUserTruckById"
        consumes:
        - "application/json"
        produces:
        - "application/json"
        parameters:
        - in: "path"
          name: "id"
          description: "Truck id"
          required: true
          type: string
        responses:
            "200":
                description: "Success"
                schema:
                    type: "object"
                    properties:
                      message: 
                        type: string
                        example: "Truck assigned successfully"
            "400":
              description: "Bad request"
              schema:
                  $ref: '#/definitions/Error'
            "500":
              description: "Internal server error"
              schema:
                  $ref: '#/definitions/Error'
        security:
        - jwt_token: []
...
securityDefinitions:
  jwt_token:
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
  Error:
    type: "object"
    properties:
      message:
        type: "string"
        description: "Error message"
...
  registrationCredentials:
    type: "object"
    properties:
      email:
        type: "string"
        example: "kyrylo@gmail.com"
      password:
        type: "string"
        example: "ve518dl3"
      role:
        type: "string"
        enum: [SHIPPER, DRIVER]
        example: "SHIPPER"
  User:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "5099803df3f4948bd2f98391"
      role:
        type: "string"
        example: "DRIVER"
      email:
        type: "string"
        example: "kyrylo@gmail.com"
      created_date:
        type: "string"
        format: "date"
        example: "2020-10-28T08:03:19.814Z"
  Truck:
    type: "object"
    properties:
      _id:
        type: "string"
        example: "5099803df3f4948bd2f98391"
      created_by:
        type: "string"
        example: "5099803df3f4948bd2f98391"
      assigned_to:
        type: "string"
        example: "5099803df3f4948bd2f98391"
      type:
        type: "string"
        enum: ["SPRINTER", "SMALL STRAIGHT", "LARGE STRAIGHT"]
        example: "SPRINTER"
      status:
        type: "string"
        enum: ["OL", "IS"]
        example: "IS"
      created_date:
        type: "string"
        format: "date"
        example: "2020-10-28T08:03:19.814Z"
...